# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Public Profile Search Prototype API
  version: 0.2.0
  description: >-
    Local-only contract for Maigret-backed cross-platform candidate discovery
    and the legacy eligibility-gated deterministic Fast Brief. Candidate
    existence is discovery evidence, not proof that accounts belong to the
    same person.
servers:
  - url: http://localhost:8800
...
paths:
...
  /v1/footprint-jobs:
    get:
      tags: [Footprint discovery]
      operationId: listFootprintHistory
      description: >-
        Returns the owner's unexpired footprint-search history grouped by exact
        normalized seed. Groups are ordered by their latest run. Profile URLs
        and platform-qualified handles share a group when they normalize to the
        same platform and handle; bare handles remain separate.
      security:
        - prototypeAuth: []
      parameters:
        - $ref: "#/components/parameters/PrototypeUser"
        - name: q
          in: query
          description: Optional case-insensitive seed identifier filter.
          schema:
            type: string
            maxLength: 160
        - name: cursor
          in: query
          description: Opaque cursor returned by the previous page.
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Newest-first page of exact-seed history groups.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FootprintHistoryGroupPage"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [Footprint discovery]
      operationId: createFootprintJob
      description: >-
        Starts a bounded Maigret-backed search for public account candidates
        from a direct profile URL, platform handle, or bare handle. A claimed
        account is returned as an unresolved candidate, not as a same-person
        conclusion.
      security:
        - prototypeAuth: []
      parameters:
        - $ref: "#/components/parameters/PrototypeUser"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFootprintJobRequest"
      responses:
        "200":
          description: Latest compatible active or usable unexpired job reused.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FootprintJob"
        "202":
          description: New discovery job accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FootprintJob"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Footprint discovery]
      operationId: clearFootprintHistory
      description: >-
        Permanently deletes a bounded batch of terminal footprint jobs owned by
        the caller. Queued and discovering jobs are preserved.
      security:
        - prototypeAuth: []
      parameters:
        - $ref: "#/components/parameters/PrototypeUser"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 50
      responses:
        "200":
          description: Bounded clear-history result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClearFootprintHistoryResponse"
        "401":
          $ref: "#/components/responses/Error"
  /v1/footprint-jobs/{job_id}:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/PrototypeUser"
    get:
      tags: [Footprint discovery]
      operationId: getFootprintJob
      security:
        - prototypeAuth: []
      responses:
        "200":
          description: Owner-visible discovery status and aggregate catalog coverage.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FootprintJob"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Footprint discovery]
      operationId: deleteFootprintJob
      security:
        - prototypeAuth: []
      responses:
        "204":
          description: Discovery data removed and a resurrection tombstone written.
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/footprint-jobs/{job_id}/history:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/PrototypeUser"
    get:
      tags: [Footprint discovery]
      operationId: listFootprintJobHistory
      description: >-
        Returns unexpired runs in the referenced job's exact normalized-seed
        group, ordered newest first.
      security:
        - prototypeAuth: []
      parameters:
        - name: cursor
          in: query
          description: Opaque cursor returned by the previous page.
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Newest-first page of runs for the exact normalized seed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FootprintHistoryRunPage"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/footprint-jobs/{job_id}/refresh:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/PrototypeUser"
    post:
      tags: [Footprint discovery]
      operationId: refreshFootprintJob
      description: >-
        Creates a new run using the source job's seed and user-selected search
        settings. Historical positives may prioritize fresh checks but are
        never copied into the new result or treated as current evidence.
      security:
        - prototypeAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Refreshed discovery job accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FootprintJob"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /v1/footprint-jobs/{job_id}/cancel:
    post:
      tags: [Footprint discovery]
      operationId: cancelFootprintJob
      description: >-
        Idempotently stops unfinished discovery work while preserving candidates
        and evidence accepted before the cancellation fence. If normal completion
        wins the race, the completed job is returned unchanged.
      security:
        - prototypeAuth: []
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/PrototypeUser"
      responses:
        "200":
          description: Durable cancelled state, or an earlier terminal state unchanged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FootprintJob"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/footprint-jobs/{job_id}/candidates:
    get:
      tags: [Footprint discovery]
      operationId: listFootprintCandidates
      description: >-
        Returns candidates discovered so far, including while catalog shards
        are still running. Relationships remain unresolved in this slice.
      security:
        - prototypeAuth: []
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/PrototypeUser"
      responses:
        "200":
          description: Current owner-visible candidate collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateList"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/footprint-jobs/{job_id}/anchor:
    post:
      tags: [Footprint discovery]
      operationId: selectFootprintAnchor
      description: >-
        Selects one exact-handle account as the known starting profile when a
        bare-handle discovery pauses on competing public name hypotheses.
        Repeating the same selection is idempotent.
      security:
        - prototypeAuth: []
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/PrototypeUser"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectFootprintAnchorRequest"
      responses:
        "200":
          description: Anchor selected and discovery resumed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelectFootprintAnchorResponse"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /v1/footprint-jobs/{job_id}/brief:
    get:
      tags: [Footprint discovery]
      operationId: getFootprintBrief
      description: >-
        Returns the current owner-visible account- or person-centric brief after
        footprint finalization. This read is scoped to footprint ownership and
        does not use the legacy self-audit eligibility gate.
      security:
        - prototypeAuth: []
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/PrototypeUser"
      responses:
        "200":
          description: Current evidence-linked footprint brief.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FootprintBrief"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/footprint-jobs/{job_id}/evidence:
    get:
      tags: [Footprint discovery]
      operationId: getFootprintEvidence
      description: >-
        Returns display-approved observations from the frozen snapshot behind
        the current footprint brief.
      security:
        - prototypeAuth: []
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/PrototypeUser"
      responses:
        "200":
          description: Evidence referenced by the current footprint report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceList"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/footprint-jobs/{job_id}/events:
    get:
      tags: [Footprint discovery]
      operationId: streamFootprintJobEvents
      security:
        - prototypeAuth: []
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/PrototypeUser"
        - name: last_event_id
          in: query
          schema:
            type: integer
            minimum: 0
        - name: Last-Event-ID
          in: header
          schema:
            type: string
      responses:
        "200":
          description: SSE stream of safe, replayable discovery JobEvent objects.
          headers:
            Cache-Control:
              schema:
                type: string
            X-Robots-Tag:
              schema:
                type: string
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    prototypeAuth:
      type: apiKey
      in: header
      name: X-Prototype-Token
      description: Local-only prototype token; never a production authentication mechanism.
    prototypeAdminAuth:
      type: apiKey
      in: header
      name: X-Prototype-Admin-Token
      description: Separate local-only operator token; never expose it to the browser.
  parameters:
    PrototypeUser:
      name: X-Prototype-User
      in: header
      required: true
      description: Local owner UUID paired with the prototype token.
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Safe machine-readable error with no profile content.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    EvidenceList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
...
    FootprintBrief:
      type: object
      additionalProperties: false
      required:
        - job_id
        - report_type
        - subject
        - summary
        - overall_identity_status
        - accounts
        - claims
        - identity_reasons
        - limitations
        - generated_at
      properties:
        job_id:
          type: string
          format: uuid
        report_type:
          type: string
          enum: [account_centric, person_centric]
        subject:
          type: string
        summary:
          type: string
        overall_identity_status:
          type: string
          enum: [confirmed, likely, unverified]
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/FootprintBriefAccount"
        claims:
          type: array
          items:
            $ref: "#/components/schemas/FootprintBriefClaim"
        identity_reasons:
          $ref: "#/components/schemas/FootprintIdentityReasons"
        narrative_sections:
          type: array
          items:
            $ref: "#/components/schemas/FootprintNarrativeSection"
        deep_story:
          anyOf:
            - $ref: "#/components/schemas/FootprintDeepStory"
            - type: "null"
        synthesis:
          $ref: "#/components/schemas/FootprintSynthesis"
        limitations:
          type: array
          items:
            type: string
        generated_at:
          type: string
          format: date-time
...
    CreateFootprintJobRequest:
      type: object
      additionalProperties: false
      required: [seed]
      dependentSchemas:
        synthesis_model:
          required: [search_mode]
          properties:
            search_mode:
              const: deep
      properties:
        seed:
          $ref: "#/components/schemas/FootprintSeed"
        search_mode:
          type: string
          enum: [quick, deep]
          default: quick
          description: >-
            Quick runs a focused account catalog and capped Exa-only professional
            search. Deep runs a broader catalog and the full configured Exa and
            GitHub search envelope before adding a source-grounded LLM story.
        synthesis_model:
          $ref: "#/components/schemas/FootprintSynthesisModel"
        locale:
          type: string
          enum: [en-US, zh-CN]
          default: en-US
        history_policy:
          type: string
          enum: [new_job, prefer_existing]
          default: new_job
          description: >-
            new_job preserves the original create behavior. prefer_existing
            reopens the latest compatible active or usable unexpired job when
            one exists, otherwise it creates a new job.
    FootprintSynthesisModel:
      type: string
      enum:
        - openai/gpt-5.6-luna
        - openai/gpt-5.4-nano
        - openai/gpt-5.4-mini
        - openai/gpt-oss-120b
        - deepseek/deepseek-v4-flash-0731
        - qwen/qwen3.5-35b-a3b
        - z-ai/glm-5.2
      description: >-
        Curated OpenRouter model used only for Deep story synthesis. If omitted,
        the server's configured synthesis gateway and model remain in effect.
    FootprintCoverage:
      type: object
      additionalProperties: false
      required: [selected, completed, claimed, available, unknown, illegal]
      properties:
        selected:
          type: integer
          minimum: 0
        completed:
          type: integer
          minimum: 0
        claimed:
          type: integer
          minimum: 0
        available:
          type: integer
          minimum: 0
        unknown:
          type: integer
          minimum: 0
        illegal:
          type: integer
          minimum: 0
...
    FootprintDeepProgress:
      type: object
      additionalProperties: false
      required: [current_phase, phase_started_at, finished_at]
      properties:
        current_phase:
          type: string
          enum:
            - queued
            - account_scan
            - awaiting_anchor
            - professional_enrichment
            - report_generation
            - finalizing
            - complete
        phase_started_at:
          type: string
          format: date-time
        finished_at:
          type: [string, "null"]
          format: date-time
    FootprintJob:
      type: object
      additionalProperties: false
      required:
        - job_id
        - status
        - exploration_status
        - deep_progress
        - seed
        - search_mode
        - synthesis_model
        - coverage
        - catalog
        - events_url
        - candidates_url
        - accepted_at
        - deadline_at
        - expires_at
        - refresh_of_job_id
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - queued
            - discovering
            - ready
            - ready_partial
            - no_candidates
            - failed
            - cancelled
        exploration_status:
          type: string
          enum: [idle, running, awaiting_anchor, completed, cancelled]
        deep_progress:
          anyOf:
            - $ref: "#/components/schemas/FootprintDeepProgress"
            - type: "null"
        seed:
          $ref: "#/components/schemas/FootprintSeedResponse"
        search_mode:
          type: [string, "null"]
          enum: [quick, deep, null]
        synthesis_model:
          anyOf:
            - $ref: "#/components/schemas/FootprintSynthesisModel"
            - type: "null"
        coverage:
          $ref: "#/components/schemas/FootprintCoverage"
        catalog:
          $ref: "#/components/schemas/FootprintCatalog"
        events_url:
          type: string
        candidates_url:
          type: string
        accepted_at:
          type: string
          format: date-time
        deadline_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        refresh_of_job_id:
          type: [string, "null"]
          format: uuid
...
    FootprintHistoryGroup:
      type: object
      additionalProperties: false
      required: [representative_job_id, seed, latest_run, run_count]
      properties:
        representative_job_id:
          type: string
          format: uuid
        seed:
          $ref: "#/components/schemas/FootprintHistorySeed"
        latest_run:
          $ref: "#/components/schemas/FootprintHistoryRun"
        run_count:
          type: integer
          minimum: 1
    FootprintHistoryGroupPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FootprintHistoryGroup"
        next_cursor:
          type: [string, "null"]
    FootprintHistoryRunPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FootprintHistoryRun"
        next_cursor:
          type: [string, "null"]
    ClearFootprintHistoryResponse:
      type: object
      additionalProperties: false
      required: [deleted_count, has_more]
      properties:
        deleted_count:
          type: integer
          minimum: 0
        has_more:
          type: boolean
    SelectFootprintAnchorRequest:
      type: object
      additionalProperties: false
      required: [candidate_id]
      properties:
        candidate_id:
          type: string
          format: uuid
...
    SelectFootprintAnchorResponse:
      type: object
      additionalProperties: false
      required: [job, selected_anchor]
      properties:
        job:
          $ref: "#/components/schemas/FootprintJob"
        selected_anchor:
          $ref: "#/components/schemas/SelectedFootprintAnchor"
...
    CandidateList:
      type: object
      additionalProperties: false
      required: [items, extracted_identifier_count]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AccountCandidate"
        extracted_identifier_count:
          type: integer
          minimum: 0
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error_code, message]
      properties:
        error_code:
          type: string
        message:
          type: string